api/requirements.txt
blinker
certifi
cffi
charset-normalizer
click
colorama
curl_cffi
Flask
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
pycryptodome
python-dotenv
requests
urllib3
Werkzeug
pytz